feat: RestaurantListViewController
