+++ solutions/devsprint-eduardo-bocato-3/DeliveryAppChallenge/Modules/Product/Home/HomeViewController.swift
@@ -10,10 +10,41 @@ import UIKit
 protocol HomeViewCategoryDisplayLogic: AnyObject {
     func displayCategoryItems(_ viewModel: Home.Category.ViewModel)
     func displaySelectedCategoryItem(_ viewModel: Home.CategorySelection.ViewModel)
-    func displayRestaurantList(_ viewModel: RestaurantList.FetchRestaurants.ViewModel)
 }
 
 class HomeViewController: UIViewController {
+    // MARK: - UI Elements
+    let scrollView: UIScrollView = {
+        
+        let scrollView = UIScrollView()
+        scrollView.translatesAutoresizingMaskIntoConstraints = false
+        return scrollView
+    }()
+
+    let stackView: UIStackView = {
+
+        let stackView = UIStackView()
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        stackView.axis = .vertical
+        stackView.alignment = .fill
+        stackView.distribution = .fill
+        stackView.spacing = 0
+        return stackView
+    }()
+
+    let addressView: AddressView = {
+
+        let addressView = AddressView()
+        addressView.translatesAutoresizingMaskIntoConstraints = false
+        return addressView
+    }()
+
+    let categoryListView: CategoryListView = {
+
+        let categoryListView = CategoryListView()
+        categoryListView.translatesAutoresizingMaskIntoConstraints = false
+        return categoryListView
+    }()
 
     //MARK: Dependencies
     private let interactor: HomeViewCategoryBusinessLogic
@@ -35,28 +66,66 @@ class HomeViewController: UIViewController {
     }
 
     override func viewDidLoad() {
+        view.backgroundColor = .white
+        
+        addSubviews()
+        configureConstraints()
         navigationController?.navigationBar.prefersLargeTitles = true
         interactor.loadCategoryItemList(.init())
-
+        setupRestaurantListView()
     }
     
     override func loadView() {
-        let customView = HomeView()
-        customView.categoryListView.didSelectItem = { [interactor] index in
+//        let customView = HomeView()
+        categoryListView.didSelectItem = { [interactor] index in
             interactor.selectCategoryItem(.init(index: index))
         }
-        self.view = customView
+//        self.view = customView
+        super.loadView()
+    }
+    
+    private func setupRestaurantListView() {
+        let child = RestaurantListAssembler().makeViewController()
+        addChild(child)
+
+        stackView.addArrangedSubview(child.view)
+        child.didMove(toParent: self)
+    }
+}
+
+extension HomeViewController {
+    func addSubviews() {
+
+        view.addSubview(scrollView)
+        scrollView.addSubview(stackView)
+
+        stackView.addArrangedSubview(addressView)
+        stackView.addArrangedSubview(categoryListView)
+    }
+
+    func configureConstraints() {
+
+        NSLayoutConstraint.activate([
+
+            scrollView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
+            scrollView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
+            scrollView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            scrollView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
+
+            stackView.leadingAnchor.constraint(equalTo: scrollView.leadingAnchor),
+            stackView.trailingAnchor.constraint(equalTo: scrollView.trailingAnchor),
+            stackView.topAnchor.constraint(equalTo: scrollView.topAnchor),
+            stackView.bottomAnchor.constraint(equalTo: scrollView.bottomAnchor),
+
+            stackView.widthAnchor.constraint(equalTo: scrollView.widthAnchor),
+        ])
     }
 }
 
 extension HomeViewController: HomeViewCategoryDisplayLogic {
-    
-    func displayRestaurantList(_ viewModel: RestaurantList.FetchRestaurants.ViewModel) {
-        
-    }
-    
+
     func displayCategoryItems(_ viewModel: Home.Category.ViewModel) {
-        (self.view as! HomeView).categoryListView.configureWith(viewModel)
+        self.categoryListView.configureWith(viewModel)
     }
 
     func displaySelectedCategoryItem(_ viewModel: Home.CategorySelection.ViewModel) {
